+++ lib/mapster.dart
@@ -4,3 +4,4 @@ export 'src/mappers/mappers.dart' hide getSource;
 export 'src/mapster/exceptions/exceptions.dart';
 export 'src/mapster/mapper_meta.dart';
 export 'src/mapster/mapster.dart';
+export 'src/mapster/to.dart';