+++ lib/src/mapster/mapper_meta.dart
@@ -1,4 +1,5 @@
 import '../mappers/mappers.dart';
+import 'to.dart';
 
 /// Callback for creating [Mapper].
 typedef MapperCreator<M extends Mapper> = M Function(MapperInput input);